+++ CHANGELOG.md
@@ -36,6 +36,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Manuel Puyol*
 
+* Fix double constant assign.
+
+    *Manuel Puyol*
+
 ## 0.0.47
 
 ### Breaking changes