+++ CHANGELOG.md
@@ -30,6 +30,8 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ## main
 
+## 0.0.48
+
 ### Misc
 
 * Expose custom cops and default config for erblint.